[refactor] DTO 생성 메서드 이름 수정 및 파일 위치 이동
- from(...)
  - 특정 객체 기반 변환
  - 여러 객체지만 한 객체가 중심
- of(...)
  - 여러 구성 요소로 조립
  - 단순 값으로 구성
- `MemberRepository`를 `repository` 디렉토리 생성 후 이동
- `@Repository` 어노테이션 불필요
