+++ src/main/java/com/hm/picplz/domain/photographer/repository/PhotoMoodRepository.java
@@ -2,13 +2,11 @@ package com.hm.picplz.domain.photographer.repository;
 
 import java.util.Optional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+
 import com.hm.picplz.domain.photographer.domain.PhotoMood;
 import com.hm.picplz.domain.photographer.domain.Photographer;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
 public interface PhotoMoodRepository extends JpaRepository<PhotoMood, Long> {
 	Optional<PhotoMood> findByPhotographerAndContent(Photographer photographer, String content);
 }
