+++ src/main/java/com/hm/picplz/domain/product/repository/ProductRepository.java
@@ -1,11 +1,11 @@
 package com.hm.picplz.domain.product.repository;
 
-import com.hm.picplz.domain.product.domain.ShootProduct;
 import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.hm.picplz.domain.product.domain.ShootProduct;
+
 public interface ProductRepository extends JpaRepository<ShootProduct, Long> {
     List<ShootProduct> findByPhotographerId(Long photographerId);
 }
